+++ src/main/java/com/mageddo/dnsproxyserver/solver/stub/addressexpression/ParseException.java
@@ -1,0 +1,12 @@
+package com.mageddo.dnsproxyserver.solver.stub.addressexpression;
+
+public class ParseException extends RuntimeException {
+
+  public ParseException(String message) {
+    super(message);
+  }
+
+  public ParseException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
